Cleaned up the new resqpy.well directory and renamed classes.
