+++ resqpy/grid_surface.py
@@ -18,7 +18,7 @@ import resqpy.olio.uuid as bu
 import resqpy.olio.vector_utilities as vec
 import resqpy.olio.xml_et as rqet
 import resqpy.surface as rqs
-import resqpy.well.well_functions as rqw
+import resqpy.well.well_utils as rqw
 import resqpy.well as rqw2
