+++ resqpy/well/WellboreFrame.py
@@ -1,4 +1,4 @@
-"""well_functions.py: resqpy well module providing trajectory, deviation survey, blocked well, wellbore frame and marker frame and md datum classes.
+"""well_utils.py: resqpy well module providing trajectory, deviation survey, blocked well, wellbore frame and marker frame and md datum classes.
 
 """
 
@@ -13,7 +13,7 @@ version = '10th November 2021'
 import logging
 
 log = logging.getLogger(__name__)
-log.debug('well_functions.py version ' + version)
+log.debug('well_utils.py version ' + version)
 
 import numpy as np
 
@@ -25,7 +25,7 @@ import resqpy.property as rqp
 from resqpy.olio.base import BaseResqpy
 from resqpy.olio.xml_namespaces import curly_namespace as ns
 
-from .well_functions import load_hdf5_array
+from .well_utils import load_hdf5_array
 
 class WellboreFrame(BaseResqpy):
     """Class for RESQML WellboreFrameRepresentation objects (supporting well log Properties)
