+++ resqpy/well/__init__.py
@@ -1,14 +1,13 @@
-__all__ = ['md_datum', 'blocked_well', 'add_wells_from_ascii_file', 'well_name', 'blocked_well_funcs', 'trajectory', 'deviation_survey', 'wellbore_frame', 'wellbore_frame_funcs']
+__all__ = ['MdDatum', 'BlockedWell', 'well_object_funcs', 'Trajectory', 'DeviationSurvey', 'WellboreFrame',
+           'WellboreMarkerFrame']
 
-from .md_datum import MdDatum
-from .wellbore_marker_frame import WellboreMarkerFrame
-from .blocked_well import BlockedWell
-from .trajectory import Trajectory
-from .deviation_survey import DeviationSurvey
-from .wellbore_frame import WellboreFrame
+from .MdDatum import MdDatum
+from .WellboreMarkerFrame import WellboreMarkerFrame
+from .BlockedWell import BlockedWell
+from .Trajectory import Trajectory
+from .DeviationSurvey import DeviationSurvey
+from .WellboreFrame import WellboreFrame
 
-from .blocked_well_funcs import add_blocked_wells_from_wellspec, add_logs_from_cellio, lookup_from_cellio
-from .add_wells_from_ascii_file import add_wells_from_ascii_file
-from .well_name import well_name
-from .wellbore_frame_funcs import add_las_to_trajectory
+from .well_object_funcs import add_wells_from_ascii_file, well_name, add_las_to_trajectory, \
+    add_blocked_wells_from_wellspec, add_logs_from_cellio, lookup_from_cellio
