+++ tests/test_model.py
@@ -9,7 +9,7 @@ import resqpy.olio.uuid as bu
 import resqpy.olio.write_hdf5 as rwh5
 import resqpy.olio.xml_et as rqet
 import resqpy.property as rqp
-import resqpy.well.well_functions as rqw
+import resqpy.well.well_utils as rqw
 
 import resqpy.well as rqw2
