+++ tests/test_well.py
@@ -6,7 +6,7 @@ from numpy.testing import assert_array_almost_equal
 from pandas.testing import assert_frame_equal
 
 import resqpy.olio.uuid as bu
-import resqpy.well.well_functions
+import resqpy.well.well_utils
 from resqpy.grid import RegularGrid
 from resqpy.model import Model
 from resqpy.property import Property
